fix(backend): support full u32 asset versions

+++ backend/src/db/migrations/007_add_asset_versions.sql
@@ -3,9 +3,9 @@
 -- version (1) by the NOT NULL defaults.
 
 ALTER TABLE assets
-    ADD COLUMN version INTEGER NOT NULL DEFAULT 1,
+    ADD COLUMN version BIGINT NOT NULL DEFAULT 1,
     ADD CONSTRAINT assets_version_positive CHECK (version >= 1);
 
 ALTER TABLE licenses
-    ADD COLUMN asset_version INTEGER NOT NULL DEFAULT 1,
+    ADD COLUMN asset_version BIGINT NOT NULL DEFAULT 1,
     ADD CONSTRAINT licenses_asset_version_positive CHECK (asset_version >= 1);
